feat: Actualizar vistas de pestañas y agregar barra de búsqueda en Tab2Page.vue y Tab3Page.vue
