+++ src/views/TabsPage.vue
@@ -10,12 +10,12 @@
 
         <ion-tab-button tab="tab2" href="/tabs/tab2">
           <ion-icon aria-hidden="true" :icon="ellipse" />
-          <ion-label>Episodios</ion-label>
+          <ion-label>Tab2</ion-label>
         </ion-tab-button>
 
         <ion-tab-button tab="tab3" href="/tabs/tab3">
           <ion-icon aria-hidden="true" :icon="square" />
-          <ion-label>Tab 3</ion-label>
+          <ion-label>Episodios</ion-label>
         </ion-tab-button>
       </ion-tab-bar>
     </ion-tabs>
